ai-service/requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.34.0
openai==1.58.1
anthropic==0.43.0
minio==7.2.12
python-dotenv==1.0.1
pydantic-settings==2.7.1
sse-starlette==2.2.1
google-genai==1.47.0
aio-pika==9.5.4
redis[hiredis]==5.2.1
PyMuPDF==1.25.3
python-docx==1.1.2
openpyxl==3.1.5
Pillow==11.1.0
eval_type_backport==0.2.2
httpx
